+++ weblate/auth/permissions.py
@@ -222,6 +222,18 @@ def check_can_edit(  # noqa: C901
         if not user.is_authenticated:
             # Signing in might help, but user still might need additional privileges
             return Denied(gettext("Sign in to save translations."))
+        if component and component.restricted:
+            if permission == "unit.review":
+                return Denied(
+                    gettext(
+                        "Insufficient privileges for approving translations in a restricted component."
+                    )
+                )
+            return Denied(
+                gettext(
+                    "Insufficient privileges for saving translations in a restricted component."
+                )
+            )
         if permission == "unit.review":
             return Denied(
                 gettext("Insufficient privileges for approving translations.")
@@ -583,7 +595,7 @@ def check_team_edit_users(
 def check_billing_view(
     user: User, permission: str, obj: Billing | Project
 ) -> bool | PermissionResult:
-    # We check Billling by hasttr to avoid importing optional Django app. To make type
+    # We check Billing by hasattr to avoid importing optional Django app. To make type
     # checker understand this, there is negative check on Project and cast in the
     # check_permission call.
     if hasattr(obj, "all_projects") and not isinstance(obj, Project):