+++ weblate/trans/views/basic.py
@@ -4,9 +4,10 @@
 from __future__ import annotations
 
 from collections import Counter
+from contextlib import suppress
 from typing import TYPE_CHECKING, Any
 
-from django.contrib.auth.decorators import login_required
+from django.contrib.auth.decorators import login_not_required, login_required
 from django.db import transaction
 from django.db.models import Q
 from django.http import HttpResponse
@@ -18,7 +19,6 @@ from django.utils.translation import gettext, ngettext
 from django.views.decorators.cache import never_cache
 from django.views.generic import RedirectView
 
-from weblate.auth.models import AuthenticatedHttpRequest
 from weblate.formats.models import EXPORTERS
 from weblate.lang.models import Language
 from weblate.trans.actions import ActionEvents
@@ -969,7 +969,7 @@ def add_languages_to_component(
 
             lang_counts[f"errors_{lang_code}"] += 1
 
-        try:
+        with suppress(FileParseError):
             # force_scan needed, see add_new_language
             if added and not component.create_translations(
                 request=request, force_scan=True
@@ -987,8 +987,6 @@ def add_languages_to_component(
                         kwargs={"path": result.get_url_path()},
                     )
                 )
-        except FileParseError:
-            pass
 
     if user.has_perm("component.edit", component):
         reset_rate_limit("language", request)
@@ -997,6 +995,7 @@ def add_languages_to_component(
 
 
 @never_cache
+@login_not_required
 def healthz(request: AuthenticatedHttpRequest) -> HttpResponse:
     """Make simple health check endpoint."""
     return HttpResponse("ok")