+++ weblate/utils/files.py
@@ -69,6 +69,7 @@ def remove_readonly(func, path, excinfo) -> None:
 
 def remove_tree(path: str | Path, ignore_errors: bool = False) -> None:
     # TODO: switch to onexc with Python >= 3.12
+    # pylint: disable-next=deprecated-argument
     shutil.rmtree(path, ignore_errors=ignore_errors, onerror=remove_readonly)